# external/memory/openapi.yaml
...
openapi: 3.1.0
info:
  title: Local Obsidian Memory Service
  summary: Local Obsidian-compatible Markdown memory backend for AI agents.
  description: The configured Markdown Vault is the source of truth. SQLite stores
    disposable search indexes, while optional Git endpoints provide explicit Vault
    versioning.
  version: 0.1.0
paths:
...
  /v1/git/init:
    post:
      tags:
      - git
      summary: Initialize Git for the Vault
      description: Initializes a Git repository in the configured Vault if one does
        not exist.
      operationId: init_git_v1_git_init_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitInitResponse'
        '400':
          description: Invalid request or Git revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested note, revision or file was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Operation conflicts with the current Vault or Git state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected database or service error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/git/status:
    get:
      tags:
      - git
      summary: Get Vault Git status
      operationId: git_status_v1_git_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitStatusResponse'
        '400':
          description: Invalid request or Git revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested note, revision or file was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Operation conflicts with the current Vault or Git state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected database or service error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/git/diff:
    get:
      tags:
      - git
      summary: Get the Markdown diff
      description: Returns the working-tree diff for Markdown files in the Vault.
      operationId: git_diff_v1_git_diff_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitDiffResponse'
        '400':
          description: Invalid request or Git revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested note, revision or file was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Operation conflicts with the current Vault or Git state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected database or service error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/git/checkpoint:
    post:
      tags:
      - git
      summary: Create a Git checkpoint
      description: Commits changed Markdown files in the Vault with the supplied message.
      operationId: checkpoint_v1_git_checkpoint_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckpointRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitCheckpointResponse'
        '400':
          description: Invalid request or Git revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested note, revision or file was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Operation conflicts with the current Vault or Git state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected database or service error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/git/history:
    get:
      tags:
      - git
      summary: List Git history
      operationId: history_v1_git_history_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitHistoryResponse'
        '400':
          description: Invalid request or Git revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested note, revision or file was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Operation conflicts with the current Vault or Git state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected database or service error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    CheckpointRequest:
      properties:
        message:
          type: string
          maxLength: 500
          minLength: 1
          title: Message
          description: Git commit message.
      type: object
      required:
      - message
      title: CheckpointRequest
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
          description: Stable machine-readable error code.
        message:
          type: string
          title: Message
          description: Human-readable error message.
      type: object
      required:
      - code
      - message
      title: ErrorDetail
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - error
      title: ErrorResponse
    GitCheckpointResponse:
      properties:
        created:
          type: boolean
          title: Created
        commit:
          anyOf:
          - type: string
          - type: 'null'
          title: Commit
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
      type: object
      required:
      - created
      title: GitCheckpointResponse
    GitCommit:
      properties:
        commit:
          type: string
          title: Commit
        timestamp:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Timestamp
        author:
          type: string
          title: Author
        message:
          type: string
          title: Message
      type: object
      required:
      - commit
      - author
      - message
      title: GitCommit
    GitDiffResponse:
      properties:
        diff:
          type: string
          title: Diff
      type: object
      required:
      - diff
      title: GitDiffResponse
    GitHistoryResponse:
      properties:
        history:
          items:
            $ref: '#/components/schemas/GitCommit'
          type: array
          title: History
      type: object
      required:
      - history
      title: GitHistoryResponse
    GitInitResponse:
      properties:
        initialized:
          type: boolean
          title: Initialized
        path:
          type: string
          title: Path
      type: object
      required:
      - initialized
      - path
      title: GitInitResponse
    GitStatusResponse:
      properties:
        initialized:
          type: boolean
          title: Initialized
        entries:
          items:
            type: string
          type: array
          title: Entries
      type: object
      required:
      - initialized
      - entries
      title: GitStatusResponse
...